Rename default AcquisitionOptimizer to GradientAcquisitionOptimizer
