+++ emukit/bayesian_optimization/local_penalization_calculator.py
@@ -6,7 +6,7 @@ from ..core.acquisition import Acquisition
 from ..core.interfaces import IDifferentiable
 from ..core.loop import CandidatePointCalculator, LoopState
 from ..bayesian_optimization.acquisitions.local_penalization import LocalPenalization
-from ..core.optimization import AcquisitionOptimizer
+from ..core.optimization import AcquisitionOptimizerBase
 from ..experimental_design.model_free.random_design import RandomDesign
 
 N_SAMPLES = 500  # Number of samples to use when estimating Lipschitz constant
@@ -21,8 +21,8 @@ class LocalPenalizationPointCalculator(CandidatePointCalculator):
     <https://arxiv.org/abs/1505.08052>`_
     """
 
-    def __init__(self, acquisition: Acquisition, acquisition_optimizer: AcquisitionOptimizer, model: IDifferentiable,
-                 parameter_space: ParameterSpace, batch_size: int):
+    def __init__(self, acquisition: Acquisition, acquisition_optimizer: AcquisitionOptimizerBase,
+                 model: IDifferentiable, parameter_space: ParameterSpace, batch_size: int):
         """
         :param acquisition: Base acquisition function to use without any penalization applied, this acquisition should
                             output positive values only.
