+++ emukit/core/optimization/__init__.py
@@ -2,7 +2,8 @@
 # SPDX-License-Identifier: Apache-2.0
 
 
-from .acquisition_optimizer import AcquisitionOptimizer  # noqa: F401
+from .acquisition_optimizer import AcquisitionOptimizerBase  # noqa: F401
+from .gradient_acquisition_optimizer import GradientAcquisitionOptimizer  # noqa: F401
 from .local_search_acquisition_optimizer import LocalSearchAcquisitionOptimizer  # noqa: F401
 from .random_search_acquisition_optimizer import RandomSearchAcquisitionOptimizer  # noqa: F401
 from .multi_source_acquisition_optimizer import MultiSourceAcquisitionOptimizer  # noqa: F401
