+++ emukit/core/optimization/multi_source_acquisition_optimizer.py
@@ -8,7 +8,7 @@ import numpy as np
 
 from .. import InformationSourceParameter, ParameterSpace
 from ..acquisition import Acquisition
-from .acquisition_optimizer import AcquisitionOptimizer, AcquisitionOptimizerBase
+from .acquisition_optimizer import AcquisitionOptimizerBase
 
 
 class MultiSourceAcquisitionOptimizer(AcquisitionOptimizerBase):
@@ -16,7 +16,7 @@ class MultiSourceAcquisitionOptimizer(AcquisitionOptimizerBase):
     Optimizes the acquisition function by finding the optimum input location at each information source, then picking
     the information source where the value of the acquisition at the optimum input location is highest.
     """
-    def __init__(self, acquisition_optimizer: AcquisitionOptimizer, space: ParameterSpace) -> None:
+    def __init__(self, acquisition_optimizer: AcquisitionOptimizerBase, space: ParameterSpace) -> None:
         """
         :param acquisition_optimizer: Optimizer to use for optimizing the acquisition once the information source
                                       has been fixed
