+++ tests/emukit/core/test_multi_source_optimizer.py
@@ -5,13 +5,13 @@ import pytest
 from emukit.core import ContinuousParameter, InformationSourceParameter, ParameterSpace
 from emukit.core.acquisition import Acquisition
 from emukit.core.loop import LoopState, SequentialPointCalculator
-from emukit.core.optimization import AcquisitionOptimizer
+from emukit.core.optimization import GradientAcquisitionOptimizer
 from emukit.core.optimization.multi_source_acquisition_optimizer import MultiSourceAcquisitionOptimizer
 
 
 @pytest.fixture
 def multi_source_optimizer():
-    mock_acquisition_optimizer = mock.create_autospec(AcquisitionOptimizer)
+    mock_acquisition_optimizer = mock.create_autospec(GradientAcquisitionOptimizer)
     mock_acquisition_optimizer.optimize.return_value = (np.array([[0.]]), None)
     space = ParameterSpace([ContinuousParameter('x', 0, 1), InformationSourceParameter(2)])
     return MultiSourceAcquisitionOptimizer(mock_acquisition_optimizer, space)
@@ -48,7 +48,7 @@ def test_multi_source_sequential_with_context():
     space = ParameterSpace([ContinuousParameter('x', 0, 1),
                             ContinuousParameter('y', 0, 1),
                             InformationSourceParameter(2)])
-    acquisition_optimizer = AcquisitionOptimizer(space)
+    acquisition_optimizer = GradientAcquisitionOptimizer(space)
     multi_source_acquisition_optimizer = MultiSourceAcquisitionOptimizer(acquisition_optimizer, space)
 
     loop_state_mock = mock.create_autospec(LoopState)
@@ -69,7 +69,7 @@ def test_multi_source_sequential_with_source_context():
     space = ParameterSpace([ContinuousParameter('x', 0, 1),
                             ContinuousParameter('y', 0, 1),
                             InformationSourceParameter(2)])
-    acquisition_optimizer = AcquisitionOptimizer(space)
+    acquisition_optimizer = GradientAcquisitionOptimizer(space)
     multi_source_acquisition_optimizer = MultiSourceAcquisitionOptimizer(acquisition_optimizer, space)
 
     loop_state_mock = mock.create_autospec(LoopState)
